init adapter module, move matrix there
