+++ src/adapter/matrix.rs
