+++ src/adapter/mod.rs
@@ -1,0 +1,2 @@
+pub mod matrix;
+pub mod pagerduty;
